+++ src/cex_platform.h
@@ -1,0 +1,16 @@
+
+#ifndef cex$platform_malloc
+#    define cex$platform_malloc malloc
+#endif
+
+#ifndef cex$platform_calloc
+#    define cex$platform_calloc calloc
+#endif
+
+#ifndef cex$platform_free
+#    define cex$platform_free free
+#endif
+
+#ifndef cex$platform_realloc
+#    define cex$platform_realloc realloc
+#endif