+++ tests/test_cexy_process.c
@@ -17,6 +17,7 @@ test$teardown_case()
     return EOK;
 }
 
+#if !defined(__EMSCRIPTEN__)
 test$case(test_src_namespace_gen)
 {
     mem$scope(tmem$, _)
@@ -161,5 +162,12 @@ test$case(test_src_namespace_update_duplicate_code)
     }
     return EOK;
 }
+
+#else
+test$case(not_supported_by_platform)
+{
+    return EOK;
+}
+#endif // #if !defined(__EMSCRIPTEN__)
 
 test$main();